+++ src/pretalx/submission/models/submission.py
@@ -94,6 +94,25 @@ class AllSubmissionManager(models.Manager):
 
 
 class Submission(LogMixin, models.Model):
+    """Submissions are, next to :class:`~pretalx.event.models.event.Event`, the central model in pretalx.
+
+    A submission, which belongs to exactly one event, can have multiple
+    speakers and a lot of other related data, such as a
+    :class:`~pretalx.submission.models.type.SubmissionType`, a
+    :class:`~pretalx.submission.models.track.Track`, multiple
+    :class:`~pretalx.submission.models.question.Answer` objects, and so on.
+
+    :param code: The unique alphanumeric identifier used to refer to a
+        submission.
+    :param state: The submission can be 'submitted', 'accepted', 'confirmed',
+        'rejected', 'withdrawn', or 'canceled'. State changes should be done via
+        the corresponding methods, like ``accept()``. The ``SubmissionStates``
+        class comes with a ``method_names`` dictionary for method lookup.
+    :param image: An image illustrating the talk or topic.
+    :param review_code: A token used in secret URLs giving read-access to the
+        submission.
+    """
+    created = models.DateTimeField(null=True, auto_now_add=True)
     code = models.CharField(max_length=16, unique=True)
     speakers = models.ManyToManyField(
         to='person.User', related_name='submissions', blank=True
@@ -181,16 +200,6 @@ class Submission(LogMixin, models.Model):
         verbose_name=_('Talk image'),
         help_text=_('Use this if you want an illustration to go with your submission.'),
     )
-    recording_url = models.CharField(
-        max_length=200, null=True, blank=True, verbose_name=_('Recording URL')
-    )
-    recording_source = models.CharField(
-        choices=(('VOC', 'media.ccc.de'),),
-        max_length=3,
-        null=True,
-        blank=True,
-        verbose_name=_('Recording Source'),
-    )
     invitation_token = models.CharField(max_length=32, default=generate_invite_code)
     review_code = models.CharField(
         max_length=32, unique=True, null=True, blank=True, default=generate_invite_code
@@ -258,12 +267,20 @@ class Submission(LogMixin, models.Model):
             return self.event.cfp.is_open or (self.event.active_review_phase and self.event.active_review_phase.speakers_can_change_submissions)
         return self.state in (SubmissionStates.ACCEPTED, SubmissionStates.CONFIRMED)
 
-    def get_duration(self):
+    def get_duration(self) -> int:
+        """Returns this submission's duration in minutes.
+
+        Falls back to the
+        :class:`~pretalx.submission.models.type.SubmissionType`'s default
+        duration if none is set on the submission."""
         if self.duration is None:
             return self.submission_type.default_duration
         return self.duration
 
     def update_duration(self):
+        """Apply the submission's duration to its currently scheduled :class:`~pretalx.schedule.models.slot.TalkSlot`.
+
+        Should be called whenever the duration changes."""
         for slot in self.event.wip_schedule.talks.filter(
             submission=self, start__isnull=False
         ):
@@ -314,6 +331,10 @@ class Submission(LogMixin, models.Model):
             )
 
     def update_talk_slots(self):
+        """Makes sure the correct amount of :class:`~pretalx.schedule.models.slot.TalkSlot` objects exists.
+
+        After an update or state change, talk slots should either be all deleted,
+        or all created, or the number of talk slots might need to be adjusted."""
         from pretalx.schedule.models import TalkSlot
 
         if self.state not in [SubmissionStates.ACCEPTED, SubmissionStates.CONFIRMED]:
@@ -345,14 +366,20 @@ class Submission(LogMixin, models.Model):
                     schedule=self.event.wip_schedule,
                 )
 
-    def make_submitted(self, person=None, force=False, orga=False):
+    def make_submitted(self, person=None, force: bool=False, orga: bool=False):
+        """Sets the submission's state to 'submitted'."""
         self._set_state(SubmissionStates.SUBMITTED, force, person=person)
 
-    def confirm(self, person=None, force=False, orga=False):
+    def confirm(self, person=None, force: bool=False, orga: bool=False):
+        """Sets the submission's state to 'confirmed'."""
         self._set_state(SubmissionStates.CONFIRMED, force, person=person)
         self.log_action('pretalx.submission.confirm', person=person, orga=orga)
 
-    def accept(self, person=None, force=False, orga=True):
+    def accept(self, person=None, force: bool=False, orga: bool=True):
+        """Sets the submission's state to 'accepted'.
+
+        Creates an acceptance :class:`~pretalx.mail.models.QueuedMail` unless
+        the submission was previously confirmed."""
         previous = self.state
         self._set_state(SubmissionStates.ACCEPTED, force, person=person)
         self.log_action('pretalx.submission.accept', person=person, orga=True)
@@ -366,7 +393,9 @@ class Submission(LogMixin, models.Model):
                     locale=self.content_locale,
                 )
 
-    def reject(self, person=None, force=False, orga=True):
+    def reject(self, person=None, force: bool=False, orga: bool=True):
+        """Sets the submission's state to 'rejected' and creates a rejection
+        :class:`~pretalx.mail.models.QueuedMail`."""
         self._set_state(SubmissionStates.REJECTED, force, person=person)
         self.log_action('pretalx.submission.reject', person=person, orga=True)
 
@@ -378,15 +407,18 @@ class Submission(LogMixin, models.Model):
                 locale=self.content_locale,
             )
 
-    def cancel(self, person=None, force=False, orga=True):
+    def cancel(self, person=None, force: bool=False, orga: bool=True):
+        """Sets the submission's state to 'canceled'."""
         self._set_state(SubmissionStates.CANCELED, force, person=person)
         self.log_action('pretalx.submission.cancel', person=person, orga=True)
 
-    def withdraw(self, person=None, force=False, orga=False):
+    def withdraw(self, person=None, force: bool=False, orga: bool=False):
+        """Sets the submission's state to 'withdrawn'."""
         self._set_state(SubmissionStates.WITHDRAWN, force, person=person)
         self.log_action('pretalx.submission.withdraw', person=person, orga=orga)
 
-    def remove(self, person=None, force=False, orga=True):
+    def remove(self, person=None, force: bool=False, orga: bool=True):
+        """Sets the submission's state to 'deleted'."""
         self._set_state(SubmissionStates.DELETED, force, person=person)
         for answer in self.answers.all():
             answer.remove(person=person, force=force)
@@ -394,6 +426,7 @@ class Submission(LogMixin, models.Model):
 
     @cached_property
     def uuid(self):
+        """A UUID5, calculated from the submission code and the instance identifier."""
         global INSTANCE_IDENTIFIER
         if not INSTANCE_IDENTIFIER:
             from pretalx.common.models.settings import GlobalSettings
@@ -435,6 +468,9 @@ class Submission(LogMixin, models.Model):
 
     @cached_property
     def slot(self):
+        """The first scheduled :class:`~pretalx.schedule.models.slot.TalkSlot` of this submission in the current :class:`~pretalx.schedule.models.schedule.Schedule`.
+
+        Note that this slot is not guaranteed to be visible."""
         return (
             self.event.current_schedule.talks.filter(submission=self).first()
             if self.event.current_schedule
@@ -443,6 +479,7 @@ class Submission(LogMixin, models.Model):
 
     @cached_property
     def display_speaker_names(self):
+        """Helper method for a consistent speaker name display."""
         return ', '.join(speaker.get_display_name() for speaker in self.speakers.all())
 
     @cached_property
@@ -486,18 +523,13 @@ class Submission(LogMixin, models.Model):
 
     @property
     def availabilities(self):
+        """The intersection of all :class:`~pretalx.schedule.models.availability.Availability` objects of all speakers of this submission."""
         from pretalx.schedule.models.availability import Availability
 
         all_availabilities = self.event.availabilities.filter(
             person__in=self.speaker_profiles
         )
         return Availability.intersection(all_availabilities)
-
-    @cached_property
-    def created(self):
-        return getattr(
-            self.logged_actions().order_by('timestamp').first(), 'timestamp', None
-        )
 
     def get_content_for_mail(self):
         order = ['title', 'abstract', 'description', 'notes', 'duration', 'content_locale', 'do_not_record', 'image']