+++ src/pretalx/submission/models/track.py
@@ -1,5 +1,7 @@
+from django.core.validators import RegexValidator
 from django.db import models
 from django.utils.text import slugify
+from django.utils.translation import ugettext_lazy as _
 from i18nfield.fields import I18nCharField
 
 from pretalx.common.mixins import LogMixin
@@ -7,11 +9,23 @@ from pretalx.common.urls import EventUrls
 
 
 class Track(LogMixin, models.Model):
+    """A track groups :class:`~pretalx.submission.models.submission.Submission`
+    objects within an :class:`~pretalx.event.models.event.Event`, e.g. by
+    topic.
+    """
     event = models.ForeignKey(
         to='event.Event', on_delete=models.PROTECT, related_name='tracks'
     )
-    name = I18nCharField(max_length=200)
-    color = models.CharField(max_length=7)
+    name = I18nCharField(
+        max_length=200,
+        verbose_name=_('Name'),
+    )
+    color = models.CharField(
+        max_length=7, verbose_name=_('Color'),
+        validators=[
+            RegexValidator(r'#([0-9A-Fa-f]{3}){1,2}'),
+        ],
+    )
 
     class urls(EventUrls):
         base = edit = '{self.event.cfp.urls.tracks}{self.pk}/'
@@ -22,5 +36,8 @@ class Track(LogMixin, models.Model):
         return str(self.name)
 
     @property
-    def slug(self):
+    def slug(self) -> str:
+        """The slug makes tracks more readable in URLs.
+
+        It consists of the ID, followed by a slugified (and, in lookups, optional) form of the track name."""
         return f'{self.id}-{slugify(self.name)}'