+++ src/tests/agenda/views/test_agenda_talks.py
@@ -80,7 +80,7 @@ def test_can_see_talk_edit_btn(
     orga_client, django_assert_num_queries, orga_user, event, slot
 ):
     slot.submission.speakers.add(orga_user)
-    with django_assert_num_queries(35):
+    with django_assert_num_queries(33):
         response = orga_client.get(slot.submission.urls.public, follow=True)
     assert response.status_code == 200
     content = response.content.decode()
@@ -250,6 +250,7 @@ def test_talk_speaker_other_submissions_only_if_visible(
 def test_talk_review_page(
     client, django_assert_num_queries, event, submission, other_submission
 ):
-    with django_assert_num_queries(17):
+    with django_assert_num_queries(22):
         response = client.get(submission.urls.review, follow=True)
     assert response.status_code == 200
+    assert submission.title in response.content.decode()